Finishing auth

+++ src/components/Campaign.js
@@ -3,24 +3,12 @@ import axios from "axios";
 import { connect } from "react-redux";
 import { deleteCampaign, getCurrentCampaign } from "../actions/actions";
 import "../index.scss";
-import { Link } from "react-router-dom";
+import { Link, Redirect } from "react-router-dom";
 
-const initialCampaign = {
-    id: 0,
-    title: "",
-    description: "",
-    urgency_level: "",
-    location: "",
-    deadline: "",
-    fund_goal: null,
-    completed: false,
-    token: ""
-}
 
 function Campaign(props) {
     console.log(props);
     let loggedIn = localStorage.getItem("token");
-    const [campaign, setCampaign] = useState(initialCampaign);
     const [donate, setDonate] = useState({});
 
     useEffect(() => {
@@ -48,16 +36,6 @@ function Campaign(props) {
             props.history.push("/");
         }, 200)
         
-        /*axios
-            .delete(`https://ptbw-sta-3.herokuapp.com/api/campaigns/${props.match.params.id}`)
-            .then(response => {
-                console.log(response)
-                props.history.replace("/")
-            })
-            .catch(error => {
-                console.log(error);
-            })*/
-        
     }
 
     const donateToCampaign = event => {
@@ -78,45 +56,31 @@ function Campaign(props) {
         setDonate({ [event.target.name]: event.target.value })
     }
 
-    
-        return (
-            <div className='news-container'>
-            <nav className='nav-bar'>
-                <Link to='/'>Home</Link>
-                <Link to='/search'>Search Campaigns</Link>
-            </nav>
-            
-            <div className="card-container">
-                <div className="Card">
-                    <h1>{props.currentCampaign.title}</h1>
-                    <h3>{props.currentCampaign.description}</h3>
-                    <div>{props.currentCampaign.urgency_level}</div>
-                    <div>{props.currentCampaign.location}</div>
-                    <div>{props.currentCampaign.deadline}</div>
-                    <div>{props.currentCampaign.fund_goal}</div>
-                    <button type="submit" onClick={routeToItem}>Update Campaign</button>
-                    <button type="submit" onClick={deleteItem}>Delete Campaign</button>
-                </div>
+        if(loggedIn) {
+            return (
+                <div className='news-container'>
+                <nav className='nav-bar'>
+                    <Link to='/'>Home</Link>
+                    <Link to='/search'>Search Campaigns</Link>
+                </nav>
                 
-            </div>
-            </div>
-        )
-     /*else {
-        return (
-            <div>
-                <h1>{campaign.title}</h1>
-                <h3>{campaign.description}</h3>
-                <div>{campaign.urgency_level}</div>
-                <div>{campaign.location}</div>
-                <div>{campaign.deadline}</div>
-                <div>{campaign.fund_goal}</div>
-                <form onSubmit={donateToCampaign}>
-                    <input onChange={onChangeDonate} type="number" placeholder="Donation" name="amount" />
-                    <button type="submit">Donate</button>
-                </form>
-            </div>
-        )
-        }*/
+                <div className="card-container">
+                    <div className="Card">
+                        <h1>{props.currentCampaign.title}</h1>
+                        <h3>{props.currentCampaign.description}</h3>
+                        <div>{props.currentCampaign.urgency_level}</div>
+                        <div>{props.currentCampaign.location}</div>
+                        <div>{props.currentCampaign.deadline}</div>
+                        <div>{props.currentCampaign.fund_goal}</div>
+                        <button type="submit" onClick={routeToItem}>Update Campaign</button>
+                        <button type="submit" onClick={deleteItem}>Delete Campaign</button>
+                    </div>
+                    
+                </div>
+                </div>
+            )
+        }
+        else return <Redirect to="/login" />
 }
 
 export default connect(state => state, { deleteCampaign: deleteCampaign, getCurrentCampaign: getCurrentCampaign })(Campaign);
